+++ packages/wdio-cucumber-framework/tests/adapter.test.js
@@ -23,10 +23,11 @@ const wdioReporter = {
     on: jest.fn()
 }
 
-const adapterFactory = (cucumberOpts = {}) => new CucumberAdapter(
+const adapterFactory = (cucumberOpts = {}, featureFlags = {}) => new CucumberAdapter(
     '0-2',
     {
         cucumberOpts,
+        featureFlags,
         beforeStep: 'beforeStep',
         afterStep: 'afterStep',
         beforeHook: 'beforeHook',
@@ -38,7 +39,7 @@ const adapterFactory = (cucumberOpts = {}) => new CucumberAdapter(
 )
 
 test('comes with a factory', async () => {
-    expect(typeof CucumberAdapterFactory.run).toBe('function')
+    expect(typeof CucumberAdapterFactory.init).toBe('function')
 })
 
 test('should properly set up cucumber', async () => {
@@ -46,6 +47,7 @@ test('should properly set up cucumber', async () => {
     adapter.registerRequiredModules = jest.fn()
     adapter.loadSpecFiles = jest.fn()
     adapter.wrapSteps = jest.fn()
+    await adapter.init()
     const result = await adapter.run()
     expect(result).toBe(0)
 
@@ -55,7 +57,7 @@ test('should properly set up cucumber', async () => {
     expect(Cucumber.setDefaultTimeout).toBeCalledWith(60000)
     expect(Cucumber.supportCodeLibraryBuilder.reset).toBeCalled()
 
-    expect(executeHooksWithArgs).toBeCalledTimes(2)
+    expect(executeHooksWithArgs).toBeCalledTimes(1)
     expect(Cucumber.PickleFilter).toBeCalled()
     expect(Cucumber.getTestCasesFromFilesystem).toBeCalled()
 })
@@ -65,6 +67,7 @@ test('should properly set up cucumber', async () => {
     adapter.registerRequiredModules = jest.fn()
     adapter.loadSpecFiles = jest.fn()
     adapter.wrapSteps = jest.fn()
+    await adapter.init()
     const result = await adapter.run()
     expect(result).toBe(0)
 
@@ -74,19 +77,35 @@ test('should properly set up cucumber', async () => {
     expect(Cucumber.setDefaultTimeout).toBeCalledWith(60000)
     expect(Cucumber.supportCodeLibraryBuilder.reset).toBeCalled()
 
-    expect(executeHooksWithArgs).toBeCalledTimes(2)
+    expect(executeHooksWithArgs).toBeCalledTimes(1)
     expect(Cucumber.PickleFilter).toBeCalled()
     expect(Cucumber.getTestCasesFromFilesystem).toBeCalled()
 })
 
-test('should throw when initialization fails', () => {
+test('should throw when initialization fails', async () => {
     const adapter = adapterFactory({ requireModule: ['@babel/register'] })
     adapter.registerRequiredModules = jest.fn()
     adapter.loadSpecFiles = jest.fn()
     adapter.wrapSteps = jest.fn()
 
     const runtimeError = new Error('boom')
-    Cucumber.Runtime.mockImplementation(() => { throw runtimeError })
+    Cucumber.PickleFilter.mockImplementationOnce(() => { throw runtimeError })
+    await expect(adapter.init()).rejects.toEqual(runtimeError)
+})
+
+test('should throw when run fails', async () => {
+    const adapter = adapterFactory({ requireModule: ['@babel/register'] })
+    adapter.registerRequiredModules = jest.fn()
+    adapter.loadSpecFiles = jest.fn()
+    adapter.wrapSteps = jest.fn()
+
+    await adapter.init()
+
+    const runtimeError = new Error('boom')
+    Cucumber.Runtime = jest.fn().mockImplementationOnce(() => ({
+        start: () => { throw runtimeError }
+    }))
+
     expect(adapter.run()).rejects.toEqual(runtimeError)
 })
 
@@ -130,6 +149,30 @@ test('loadSpecFiles', () => {
     expect(mockery.enable).toBeCalledTimes(1)
     expect(mockery.registerMock).toBeCalledTimes(1)
     expect(mockery.disable).toBeCalledTimes(1)
+})
+
+describe('hasTests', () => {
+    test('should return false if there are no tests', async () => {
+        const adapter = adapterFactory()
+        adapter.loadSpecFiles = jest.fn()
+        await adapter.init()
+        expect(adapter.hasTests()).toBe(false)
+    })
+
+    test('should return true if the feature is disabled', async () => {
+        const adapter = adapterFactory(undefined, { specFiltering: false })
+        adapter.loadSpecFiles = jest.fn()
+        await adapter.init()
+        expect(adapter.hasTests()).toBe(true)
+    })
+
+    test('should return true if there are tests', async () => {
+        Cucumber.getTestCasesFromFilesystem.mockImplementationOnce(() => ['/foo/bar.js'])
+        const adapter = adapterFactory()
+        adapter.loadSpecFiles = jest.fn()
+        await adapter.init()
+        expect(adapter.hasTests()).toBe(true)
+    })
 })
 
 describe('wrapSteps', () => {