+++ packages/wdio-mocha-framework/tests/adapter.test.js
@@ -1,4 +1,5 @@
 import path from 'path'
+import Mocha from 'mocha'
 import logger from '@wdio/logger'
 import { runTestInFiberContext, executeHooksWithArgs } from '@wdio/utils'
 
@@ -15,27 +16,30 @@ const wdioReporter = {
     emit: jest.fn(),
     on: jest.fn()
 }
+const adapterFactory = (config) => new MochaAdapter(
+    '0-2',
+    { featureFlags: {}, ...config },
+    ['/foo/bar.test.js'],
+    { browserName: 'chrome' },
+    wdioReporter
+)
 
 test('comes with a factory', async () => {
-    expect(typeof MochaAdapterFactory.run).toBe('function')
-    const result = await MochaAdapterFactory.run(
+    expect(typeof MochaAdapterFactory.init).toBe('function')
+    const instance = await MochaAdapterFactory.init(
         '0-2',
-        {},
+        { featureFlags: {} },
         ['/foo/bar.test.js'],
         { browserName: 'chrome' },
         wdioReporter
     )
+    const result = await instance.run()
     expect(result).toBe(0)
 })
 
 test('should properly set up mocha', async () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
+    await adapter.init()
     const result = await adapter.run()
     expect(result).toBe(0)
 
@@ -43,7 +47,7 @@ test('should properly set up mocha', async () => {
     expect(adapter.mocha.reporter).toBeCalled()
     expect(adapter.mocha.fullTrace).toBeCalled()
     expect(adapter.mocha.run).toBeCalled()
-    expect(executeHooksWithArgs.mock.calls).toHaveLength(2)
+    expect(executeHooksWithArgs.mock.calls).toHaveLength(1)
     expect(adapter.mocha.runner.on.mock.calls).toHaveLength(Object.keys(EVENTS).length)
     expect(adapter.mocha.runner.suite.beforeAll).toBeCalled()
     expect(adapter.mocha.runner.suite.beforeEach).not.toBeCalled()
@@ -54,37 +58,21 @@ test('should properly set up mocha', async () => {
 })
 
 test('should return amount of errors', async () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        { mochaOpts: { mockFailureCount: 42 } },
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory({ mochaOpts: { mockFailureCount: 42 } })
+    await adapter.init()
     const result = await adapter.run()
     expect(result).toBe(42)
 })
 
 test('should throw runtime error if spec is invalid', async () => {
     const runtimeError = new Error('Uuups')
-    const adapter = new MochaAdapter(
-        '0-2',
-        { mochaOpts: { mockRuntimeError: runtimeError } },
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory({ mochaOpts: { mockRuntimeError: runtimeError } })
+    await adapter.init()
     await expect(adapter.run()).rejects.toEqual(runtimeError)
 })
 
 test('options', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
     adapter.requireExternalModules = jest.fn()
     adapter.options({
         require: 'foo/bar.js',
@@ -95,13 +83,7 @@ test('options', () => {
 
 test('preRequire', () => {
     const mochaOpts = { foo: 'bar', ui: 'tdd' }
-    const adapter = new MochaAdapter(
-        '0-2',
-        { mochaOpts, beforeHook: 'beforeHook123', afterHook: 'afterHook123', beforeTest: 'beforeTest234', afterTest: 'afterTest234' },
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory({ mochaOpts, beforeHook: 'beforeHook123', afterHook: 'afterHook123', beforeTest: 'beforeTest234', afterTest: 'afterTest234' })
     adapter.preRequire('context', 'file', 'mocha')
     expect(runTestInFiberContext).toBeCalledWith(false, 'beforeHook123', expect.any(Function), 'afterHook123', expect.any(Function), 'suiteSetup', '0-2')
     expect(runTestInFiberContext).toBeCalledWith(false, 'beforeHook123', expect.any(Function), 'afterHook123', expect.any(Function), 'setup', '0-2')
@@ -116,13 +98,7 @@ test('preRequire', () => {
 
 test('custom ui', () => {
     const mochaOpts = { ui: 'custom-qunit' }
-    const adapter = new MochaAdapter(
-        '0-2',
-        { mochaOpts },
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory({ mochaOpts })
     adapter.preRequire('context', 'file', 'mocha')
     expect(runTestInFiberContext).toBeCalledWith(false, undefined, expect.any(Function), undefined, expect.any(Function), 'after', '0-2')
     expect(runTestInFiberContext).toBeCalledWith(false, undefined, expect.any(Function), undefined, expect.any(Function), 'afterEach', '0-2')
@@ -132,13 +108,7 @@ test('custom ui', () => {
 
 test('wrapHook if successful', async () => {
     const config = { beforeAll: 'somehook' }
-    const adapter = new MochaAdapter(
-        '0-2',
-        config,
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory(config)
     const wrappedHook = adapter.wrapHook('beforeAll')
 
     executeHooksWithArgs.mockImplementation((...args) => Promise.resolve(args))
@@ -149,13 +119,7 @@ test('wrapHook if successful', async () => {
 
 test('wrapHook if failing', async () => {
     const config = { beforeAll: 'somehook' }
-    const adapter = new MochaAdapter(
-        '0-2',
-        config,
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory(config)
     const wrappedHook = adapter.wrapHook('beforeAll')
 
     executeHooksWithArgs.mockImplementation(() => Promise.reject(new Error('uuuups')))
@@ -166,13 +130,8 @@ test('wrapHook if failing', async () => {
 })
 
 test('prepareMessage', async () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
+    await adapter.init()
     await adapter.run()
 
     adapter.lastError = new Error('uuups')
@@ -188,13 +147,7 @@ test('prepareMessage', async () => {
 })
 
 test('formatMessage', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
 
     let params = { type: 'foobar' }
     let message = adapter.formatMessage(params)
@@ -246,38 +199,20 @@ test('formatMessage', () => {
 })
 
 test('requireExternalModules', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
     adapter.requireExternalModules(['/foo/bar.js', null, './bar/foo.js'], { myContext: 123 })
     expect(loadModule).toBeCalledWith('/foo/bar.js', { myContext: 123 })
     expect(loadModule).toBeCalledWith(path.resolve(__dirname, '..', '..', '..', 'bar', 'foo.js'), { myContext: 123 })
 })
 
 test('emit does not emit anything on root level', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
     adapter.emit(null, { root: true })
     expect(wdioReporter.emit).not.toBeCalled()
 })
 
 test('emit properly reports to reporter', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
     adapter.getUID = () => 123
     adapter.emit(
         'suite:start',
@@ -292,13 +227,7 @@ test('emit properly reports to reporter', () => {
 })
 
 test('emits hook errors as hook:end', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
     adapter.getUID = () => 123
     adapter.emit(
         'test:fail',
@@ -311,13 +240,7 @@ test('emits hook errors as hook:end', () => {
 })
 
 test('getUID', () => {
-    const adapter = new MochaAdapter(
-        '0-2',
-        {},
-        ['/foo/bar.test.js'],
-        { browserName: 'chrome' },
-        wdioReporter
-    )
+    const adapter = adapterFactory()
 
     // disabling indent eslint rule for better visibility
     /*eslint-disable indent */
@@ -377,7 +300,65 @@ test('getUID', () => {
     /*eslint-enable indent */
 })
 
+describe('loadFiles', () => {
+    test('should do nothing if feature is not enabled', () => {
+        const adapter = adapterFactory()
+        adapter._hasTests = null
+        expect(adapter._loadFiles({})).toBe(false)
+        expect(adapter._hasTests).toBe(null)
+    })
+
+    test('should set _hasTests to true if there are tests to run', () => {
+        const adapter = adapterFactory({ featureFlags: { specFiltering: true } })
+        adapter._hasTests = null
+        adapter.mocha = {
+            loadFiles: jest.fn(),
+            suite: 1 // mochaRunner.total
+        }
+        adapter._loadFiles({})
+        expect(adapter._hasTests).toBe(true)
+    })
+
+    test('should set _hasTests to false if there no tests to run', () => {
+        const adapter = adapterFactory({ featureFlags: { specFiltering: true } })
+        adapter._hasTests = null
+        adapter.mocha = {
+            loadFiles: jest.fn(),
+            options: { grep: 'regexp foo' },
+            suite: 0 // mochaRunner.total
+        }
+        adapter._loadFiles({ grep: 'foo', invert: 'invert' })
+        expect(Mocha.Runner.mock.results[0].value.grep).toBeCalledWith('regexp foo', 'invert')
+        expect(adapter._hasTests).toBe(false)
+    })
+
+    test('should not fail on exception', () => {
+        const adapter = adapterFactory({ featureFlags: { specFiltering: true } })
+        adapter._hasTests = null
+        adapter.mocha = {
+            loadFiles: jest.fn().mockImplementation(() => { throw new Error('foo') }),
+        }
+        adapter._loadFiles({})
+        expect(adapter.mocha.loadFiles).toBeCalled()
+        expect(adapter._hasTests).toBe(null)
+    })
+})
+
+describe('hasTests', () => {
+    test('should return true if feature is not enabled', () => {
+        const adapter = adapterFactory()
+        adapter._hasTests = 'foobar'
+        expect(adapter.hasTests()).toBe(true)
+    })
+    test('should return _hasTests if feature is enabled', () => {
+        const adapter = adapterFactory({ featureFlags: { specFiltering: true } })
+        adapter._hasTests = 'foobar'
+        expect(adapter.hasTests()).toBe('foobar')
+    })
+})
+
 afterEach(() => {
+    Mocha.Runner.mockClear()
     runTestInFiberContext.mockReset()
     executeHooksWithArgs.mockReset()
 })