+++ packages/webdriver/src/utils.js
@@ -10,12 +10,6 @@ import command from './command'
 
 const log = logger('webdriver')
 
-const MOBILE_BROWSER_NAMES = ['ipad', 'iphone', 'android']
-const MOBILE_CAPABILITIES = [
-    'appium-version', 'appiumVersion', 'device-type', 'deviceType',
-    'device-orientation', 'deviceOrientation', 'deviceName'
-]
-
 const BROWSER_DRIVER_ERRORS = [
     'unknown command: wd/hub/session', // chromedriver
     'HTTP method not allowed', // geckodriver
@@ -186,146 +180,6 @@ export function getPrototype ({ isW3C, isChrome, isMobile, isSauce, isSeleniumSt
 }
 
 /**
- * check if session is based on W3C protocol based on the /session response
- * @param  {Object}  capabilities  caps of session response
- * @return {Boolean}               true if W3C (browser)
- */
-export function isW3C (capabilities) {
-    /**
-     * JSONWire protocol doesn't return a property `capabilities`.
-     * Also check for Appium response as it is using JSONWire protocol for most of the part.
-     */
-    if (!capabilities) {
-        return false
-    }
-
-    /**
-     * assume session to be a WebDriver session when
-     * - capabilities are returned
-     *   (https://w3c.github.io/webdriver/#dfn-new-sessions)
-     * - it is an Appium session (since Appium is full W3C compliant)
-     */
-    const isAppium = capabilities.automationName || capabilities.deviceName || (capabilities.appiumVersion)
-    const hasW3CCaps = (
-        capabilities.platformName &&
-        capabilities.browserVersion &&
-        /**
-         * local safari and BrowserStack don't provide platformVersion therefor
-         * check also if setWindowRect is provided
-         */
-        (capabilities.platformVersion || Object.prototype.hasOwnProperty.call(capabilities, 'setWindowRect'))
-    )
-    return Boolean(hasW3CCaps || isAppium)
-}
-
-/**
- * check if session is run by Chromedriver
- * @param  {Object}  capabilities  caps of session response
- * @return {Boolean}               true if run by Chromedriver
- */
-export function isChrome (caps) {
-    return (
-        Boolean(caps.chrome) ||
-        Boolean(caps['goog:chromeOptions'])
-    )
-}
-
-/**
- * check if current platform is mobile device
- *
- * @param  {Object}  caps  capabilities
- * @return {Boolean}       true if platform is mobile device
- */
-export function isMobile (caps) {
-    const browserName = (caps.browserName || '').toLowerCase()
-
-    /**
-     * we have mobile caps if
-     */
-    return Boolean(
-        /**
-         * capabilities contain mobile only specific capabilities
-         */
-        Object.keys(caps).find((cap) => MOBILE_CAPABILITIES.includes(cap)) ||
-        /**
-         * browserName is empty (and eventually app is defined)
-         */
-        caps.browserName === '' ||
-        /**
-         * browserName is a mobile browser
-         */
-        MOBILE_BROWSER_NAMES.includes(browserName)
-    )
-}
-
-/**
- * check if session is run on iOS device
- * @param  {Object}  capabilities  caps of session response
- * @return {Boolean}               true if run on iOS device
- */
-export function isIOS (caps) {
-    return Boolean(
-        (caps.platformName && caps.platformName.match(/iOS/i)) ||
-        (caps.deviceName && caps.deviceName.match(/(iPad|iPhone)/i))
-    )
-}
-
-/**
- * check if session is run on Android device
- * @param  {Object}  capabilities  caps of session response
- * @return {Boolean}               true if run on Android device
- */
-export function isAndroid (caps) {
-    return Boolean(
-        (caps.platformName && caps.platformName.match(/Android/i)) ||
-        (caps.browserName && caps.browserName.match(/Android/i))
-    )
-}
-
-/**
- * detects if session is run on Sauce with extended debugging enabled
- * @param  {string}  hostname     hostname of session request
- * @param  {object}  capabilities session capabilities
- * @return {Boolean}              true if session is running on Sauce with extended debugging enabled
- */
-export function isSauce (hostname, caps) {
-    return Boolean(
-        caps.extendedDebugging ||
-        (
-            caps['sauce:options'] &&
-            caps['sauce:options'].extendedDebugging
-        )
-    )
-}
-
-/**
- * detects if session is run using Selenium Standalone server
- * @param  {object}  capabilities session capabilities
- * @return {Boolean}              true if session is run with Selenium Standalone Server
- */
-export function isSeleniumStandalone (caps) {
-    return Boolean(caps['webdriver.remote.sessionid'])
-}
-
-/**
- * returns information about the environment
- * @param  {Object}  hostname      name of the host to run the session against
- * @param  {Object}  capabilities  caps of session response
- * @return {Object}                object with environment flags
- */
-export function environmentDetector ({ hostname, capabilities, requestedCapabilities }) {
-    return {
-        isW3C: isW3C(capabilities),
-        isChrome: isChrome(capabilities),
-        isMobile: isMobile(capabilities),
-        isIOS: isIOS(capabilities),
-        isAndroid: isAndroid(capabilities),
-        isSauce: isSauce(hostname, requestedCapabilities.w3cCaps.alwaysMatch),
-        isSeleniumStandalone: isSeleniumStandalone(capabilities)
-    }
-}
-
-/**
  * helper method to determine the error from webdriver response
  * @param  {Object} body body object
  * @return {Object} error
@@ -364,7 +218,7 @@ export class CustomRequestError extends Error {
  * @param  {Object} options   driver instance or option object containing these flags
  * @return {Object}           prototype object
  */
-export function getEnvironmentVars({ isW3C, isMobile, isIOS, isAndroid, isChrome, isSauce }) {
+export function getEnvironmentVars({ isW3C, isMobile, isIOS, isAndroid, isChrome, isSauce, isSeleniumStandalone }) {
     return {
         isW3C: { value: isW3C },
         isMobile: { value: isMobile },
@@ -404,7 +258,8 @@ export function setupDirectConnect(params) {
 export const getSessionError = (err) => {
     // browser driver / service is not started
     if (err.code === 'ECONNREFUSED') {
-        return `Unable to connect to "${err.address}:${err.port}", make sure browser driver is running on that address.`
+        return `Unable to connect to "${err.address}:${err.port}", make sure browser driver is running on that address.` +
+            '\nIf you use services like chromedriver see initialiseServices logs above or in wdio.log file.'
     }
 
     if (!err.message) {