+++ tests/helpers/shared-store.conf.js
@@ -1,10 +1,8 @@
 const { config } = require('./config')
 
-const conf = Object.assign({}, config, {
+exports.config = Object.assign({}, config, {
     beforeSuite () {
         browser.sharedStore.set(browser.sessionId, browser.capabilities)
     },
+    services: [...config.services, 'shared-store']
 })
-conf.services.push('shared-store')
-
-exports.config = conf